Use the new component in the settings page
