+++ apps/extension/src/routes/page/paths.ts
@@ -6,7 +6,7 @@ export enum PagePath {
   IMPORT_SEED_PHRASE = '/welcome/import',
   ONBOARDING_SUCCESS = '/welcome/success',
   SET_PASSWORD = '/welcome/set-password',
-  SET_RPC_ENDPOINT = '/welcome/set-rpc-endpoint',
+  SET_GRPC_ENDPOINT = '/welcome/set-grpc-endpoint',
   RESTORE_PASSWORD = '/restore-password',
   RESTORE_PASSWORD_INDEX = '/restore-password/',
   RESTORE_PASSWORD_SET_PASSWORD = '/restore-password/set-password',
